Add editTask tests ok
