+++ src/index.js
@@ -1,65 +1,5 @@
 import './index.css';
-import {
-  addTask,
-  deleteTask,
-  editTask,
-  updateTasks,
-  toggleCompleted,
-  filterTasks,
-} from './methods.js';
-
-const tasksContainer = document.getElementById('tasks-container');
-
-export default function renderTasks() {
-  const tasks = updateTasks();
-  tasksContainer.innerHTML = '';
-  tasks.forEach((task, idx) => {
-    const listItem = document.createElement('li');
-    const checkbox = document.createElement('input');
-    const input = document.createElement('input');
-    const dotsButton = document.createElement('input');
-
-    checkbox.setAttribute('type', 'checkbox');
-    checkbox.setAttribute('tabindex', '0');
-    checkbox.setAttribute('alt', 'Check!');
-    checkbox.checked = task.completed;
-
-    input.setAttribute('maxlength', '255');
-    input.value = task.description;
-
-    const attrs = {
-      type: 'button',
-      tabindex: '-1',
-      value: '',
-      title: 'click and sostain for rearrange',
-      class: 'input-btn-dots',
-    };
-
-    Object.entries(attrs).forEach(([key, value]) => {
-      dotsButton.setAttribute(key, value);
-    });
-
-    listItem.appendChild(checkbox);
-    listItem.appendChild(input);
-    listItem.appendChild(dotsButton);
-
-    tasksContainer.appendChild(listItem);
-
-    input.addEventListener('input', () => {
-      editTask(idx + 1, input.value);
-    });
-
-    checkbox.addEventListener('change', () => {
-      toggleCompleted(idx + 1);
-      renderTasks();
-    });
-
-    dotsButton.addEventListener('click', () => {
-      deleteTask(idx + 1);
-      renderTasks();
-    });
-  });
-}
+import { addTask, filterTasks, renderTasks } from './methods.js';
 
 const clearBtn = document.getElementById('clear-btn');
 clearBtn.addEventListener('click', () => {
@@ -70,7 +10,8 @@ clearBtn.addEventListener('click', () => {
 const inputList = document.getElementById('input-list');
 const form = document.getElementById('form-input');
 
-form.addEventListener('submit', () => {
+form.addEventListener('submit', (e) => {
+  e.preventDefault();
   const description = inputList.value;
   if (description !== '') {
     addTask(description);
@@ -80,5 +21,3 @@ form.addEventListener('submit', () => {
 });
 
 renderTasks();
-
-export { renderTasks };
