+++ src/methods.js
@@ -49,13 +49,73 @@ const toggleCompleted = (id) => {
 
 const filterTasks = () => {
   const filteredTasks = tasks.filter((task) => task.completed === false);
-  filteredTasks.forEach((task, index) => { task.id = index + 1; });
+  filteredTasks.forEach((task, index) => {
+    task.id = index + 1;
+  });
   tasks = filteredTasks;
   tasksObj.setTasks(filteredTasks);
   tasks = tasksObj.getTasks();
   updateTasks();
 };
 
+function renderTasks() {
+  const tasksContainer = document.getElementById('tasks-container');
+  const tasks = updateTasks();
+  tasksContainer.innerHTML = '';
+  tasks.forEach((task, idx) => {
+    const listItem = document.createElement('li');
+    const checkbox = document.createElement('input');
+    const input = document.createElement('input');
+    const dotsButton = document.createElement('input');
+
+    checkbox.setAttribute('type', 'checkbox');
+    checkbox.setAttribute('tabindex', '0');
+    checkbox.setAttribute('alt', 'Check!');
+    checkbox.checked = task.completed;
+
+    input.setAttribute('maxlength', '255');
+    input.value = task.description;
+
+    const attrs = {
+      type: 'button',
+      tabindex: '-1',
+      value: '',
+      title: 'click and sostain for rearrange',
+      class: 'input-btn-dots',
+    };
+
+    Object.entries(attrs).forEach(([key, value]) => {
+      dotsButton.setAttribute(key, value);
+    });
+
+    listItem.appendChild(checkbox);
+    listItem.appendChild(input);
+    listItem.appendChild(dotsButton);
+
+    tasksContainer.appendChild(listItem);
+
+    input.addEventListener('input', () => {
+      editTask(idx + 1, input.value);
+    });
+
+    checkbox.addEventListener('change', () => {
+      toggleCompleted(idx + 1);
+      renderTasks();
+    });
+
+    dotsButton.addEventListener('click', () => {
+      deleteTask(idx + 1);
+      renderTasks();
+    });
+  });
+}
+
 export {
-  updateTasks, addTask, deleteTask, editTask, toggleCompleted, filterTasks,
+  updateTasks,
+  addTask,
+  deleteTask,
+  editTask,
+  toggleCompleted,
+  filterTasks,
+  renderTasks,
 };
